pyproject.toml
[project]
name = "bragi"
version = "0.1.0"
description = "Self-hosted, OpenAI-compatible STT/TTS server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.18",
    "soundfile>=0.13.0",
    "soxr>=0.5.0",
    "numpy>=2.0.0",
    "lameenc>=1.8.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "aiosqlite>=0.20.0",
    "faster-whisper>=1.1.0",
    "kokoro>=0.9.4",
]

[project.optional-dependencies]
vosk = ["vosk>=0.3.40"]
funasr = ["funasr>=1.0.0"]
moonshine = ["useful-moonshine-onnx>=20251120"]
speechbrain = ["speechbrain>=1.0.0"]
parakeet = ["nemo_toolkit[asr]>=2.0.0"]
piper = ["piper-tts>=1.2.0"]
coqui = ["coqui-tts>=0.27.0"]
f5tts = ["f5-tts>=0.3.0"]
qwen-tts = ["qwen-tts>=0.1.0"]
all = [
    "vosk>=0.3.40",
    "funasr>=1.0.0",
    "useful-moonshine-onnx>=20251120",
    "speechbrain>=1.0.0",
    "nemo_toolkit[asr]>=2.0.0",
    "piper-tts>=1.2.0",
    "coqui-tts>=0.27.0",
    "f5-tts>=0.3.0",
    "qwen-tts>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bragi"]
